quadruped_ctrl/requirements.txt
mujoco==3.4.0
numpy==2.4.1
gymnasium==1.2.3
pyyaml==6.0.3
casadi==3.7.0
matplotlib==3.10.8
scipy==1.15.1
jax
jaxlib
tqdm

